+styles file & add title to home view
