+++ lib/Features/home/presentation/views/widgets/featured_list_view.dart
@@ -1,0 +1,22 @@
+import 'package:booklo/Features/home/presentation/views/widgets/custom_list_view_item.dart';
+import 'package:flutter/material.dart';
+
+class FeaturedBooksListView extends StatelessWidget {
+  const FeaturedBooksListView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: MediaQuery.of(context).size.height * 0.3,
+      child: ListView.builder(
+        scrollDirection: Axis.horizontal,
+        itemBuilder: (context, index) {
+          return const Padding(
+            padding: EdgeInsets.symmetric(horizontal: 8),
+            child: FeaturedListViewItem(),
+          );
+        },
+      ),
+    );
+  }
+}
