+++ lib/Features/home/presentation/views/widgets/home_view_body.dart
@@ -1,5 +1,6 @@
 import 'package:booklo/Features/home/presentation/views/widgets/custom_app_bar.dart';
-import 'package:booklo/Features/home/presentation/views/widgets/custom_list_view_item.dart';
+import 'package:booklo/Features/home/presentation/views/widgets/featured_list_view.dart';
+import 'package:booklo/core/utils/styles.dart';
 import 'package:flutter/material.dart';
 
 class HomeViewBody extends StatelessWidget {
@@ -7,31 +8,20 @@ class HomeViewBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return const Column(
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         CustomAppBar(),
         FeaturedBooksListView(),
+        SizedBox(height: 20),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 26),
+          child: Text(
+            'Best Seller',
+            style: Styles.titleMedium,
+          ),
+        ),
       ],
-    );
-  }
-}
-
-class FeaturedBooksListView extends StatelessWidget {
-  const FeaturedBooksListView({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      height: MediaQuery.of(context).size.height * 0.3,
-      child: ListView.builder(
-        scrollDirection: Axis.horizontal,
-        itemBuilder: (context, index) {
-          return const Padding(
-            padding: EdgeInsets.symmetric(horizontal: 8),
-            child: FeaturedListViewItem(),
-          );
-        },
-      ),
     );
   }
 }
