+++ lib/core/utils/styles.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+abstract class Styles {
+  static const titleMedium = TextStyle(
+    fontSize: 19,
+    fontWeight: FontWeight.w600,
+  );
+}
